- Added the C99 printf size specifiers 't' (ptrdiff_t) and 'z' (size_t) to   FString::Format() so that I can fix all the problem printf strings that a   64-bit GCC compile finds.
SVN r968 (trunk)

+++ src/f_finale.cpp
@@ -961,7 +961,7 @@ void F_BunnyScroll (void)
 			laststage = stage;
 		}
 
-		sprintf (name, "END%i", stage);
+		sprintf (name, "END%d", (int)stage);
 		screen->DrawTexture (TexMan(name), (320-13*8)/2, (200-8*8)/2, DTA_320x200, true, TAG_DONE);
 	}
 }
